fix: resolve SonarQube issues - extract shared utils, use top-level await
